--- validator/regexp_test.go
@@ -1,21 +1,0 @@
-package validator_test
-
-import (
-	"regexp"
-	"testing"
-
-	"fyne.io/fyne/validator"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestValidator_Regexp(t *testing.T) {
-	r := validator.Regexp{Reason: "Input is not valid"}
-
-	assert.NoError(t, r.Validate("nothing")) // Nothing to validate with, same as no validator.
-
-	r.Regexp = regexp.MustCompile(`^\d{2}-\w{4}$`)
-
-	assert.NoError(t, r.Validate("10-fyne"))
-	assert.EqualError(t, r.Validate("100-fynedesk"), r.Reason)
-}
